requirements.txt
aiofiles==0.8.0
amqp==5.1.1
anyio==3.6.1
aredis==1.1.8
async-timeout==4.0.2
asynch==0.1.9
backports.zoneinfo==0.2.1
billiard==3.6.4.0
ciso8601==2.2.0
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
clickhouse-cityhash==1.0.2.3
Deprecated==1.2.13
fastapi==0.79.0
h11==0.13.0
idna==3.3
kombu==5.2.4
leb128==1.0.4
loguru==0.6.0
lz4==4.0.1
numpy==1.23.1
packaging==21.3
pandas==1.4.3
prompt-toolkit==3.0.30
pydantic==1.9.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
python-multipart==0.0.5
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
redis==4.3.4
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
typing_extensions==4.3.0
tzdata==2022.1
tzlocal==4.2
ujson==5.4.0
uvicorn==0.18.2
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
zstd==1.5.2.5
